refactor: IndexFileManager 생성
- index 파일 관련 처리를 전담하는 IndexFileManager 클래스 생성
