+++ src/main/kotlin/geet/commands/porcelain/geetAdd.kt
@@ -1,7 +1,7 @@
 package geet.commands.porcelain
 
-import geet.exception.BadRequest
-import geet.util.createGeetObjectWithFile
+import geet.exceptions.BadRequest
+import geet.utils.createGeetObjectWithFile
 import java.io.File
 
 fun geetAdd(commandLines: Array<String>): Unit {
