+++ src/main/kotlin/geet/exceptions/BadRequest.kt
@@ -1,4 +1,4 @@
-package geet.exception
+package geet.exceptions
 
 class BadRequest(override val message: String): Exception() {
     val statusCode: Int = 400
