+++ src/main/kotlin/geet/exceptions/NotModifiedObject.kt
@@ -1,4 +1,4 @@
-package geet.exception
+package geet.exceptions
 
 class NotModifiedObject(
     override val message: String
