--- src/main/kotlin/geet/util/const.kt
@@ -1,5 +1,0 @@
-package geet.util
-
-import java.security.MessageDigest
-
-val messageDigest: MessageDigest = java.security.MessageDigest.getInstance("SHA-1")
